+++ game/world/managers/maps/GridManager.py
@@ -3,7 +3,9 @@ from __future__ import annotations
 import math
 import time
 
+from game.world.managers.maps.Cell import Cell
 from utils.ConfigManager import config
+from utils.Logger import Logger
 from utils.constants.MiscCodes import ObjectTypeIds
 
 TOLERANCE = 0.00001
@@ -18,15 +20,15 @@ class GridManager:
         self.cells: dict[str, Cell] = {}
         self.active_cell_callback = active_cell_callback
 
-    def get_create_cell(self, world_object):
-        cell_key = GridManager.get_cell_key(world_object.location.x, world_object.location.y, world_object.map_)
-        cell = self.cells.get(cell_key)
-        if not cell:
-            min_x, min_y, max_x, max_y = GridManager.generate_coord_data(world_object.location.x,
-                                                                         world_object.location.y)
-            cell = Cell(min_x, min_y, max_x, max_y, world_object.map_)
-            self.cells[cell.key] = cell
-        return cell
+    def spawn_object(self, world_object_spawn=None, world_object_instance=None):
+        location = world_object_spawn.location if world_object_spawn else world_object_instance.location
+        map_ = world_object_spawn.map_ if world_object_spawn else world_object_instance.map_
+        if world_object_instance:
+            self.add_world_object(world_object_instance)
+        if world_object_spawn:
+            self.add_world_object_spawn(world_object_spawn)
+        if not world_object_spawn and not world_object_instance:
+            Logger.warning(f'Spawn object called with None arguments.')
 
     def update_object(self, world_object, old_grid_manager, has_changes=False, has_inventory_changes=False):
         source_cell_key = world_object.current_cell
@@ -37,13 +39,13 @@ class GridManager:
             # Remove from old location.
             old_grid_manager.remove_object(world_object)
             # Add to new location.
-            self.add_object(world_object)
+            self.add_world_object(world_object)
         # Handle cell change within the same map.
         elif current_cell_key != source_cell_key:
             # Remove from old location and Add to new location.
             if source_cell_key:
                 self.remove_object(world_object, update_players=False)
-            self.add_object(world_object, update_players=False)
+            self.add_world_object(world_object, update_players=False)
             # Update old location surroundings, even if in the same grid, both cells quadrants might not see each other.
             affected_cells = self.update_players(source_cell_key)
             # Update new location surroundings, excluding intersecting cells from previous call.
@@ -58,9 +60,13 @@ class GridManager:
         if old_grid_manager and old_grid_manager != self or current_cell_key != source_cell_key:
             world_object.on_cell_change()
 
-    def add_object(self, world_object, update_players=True):
-        cell = self.get_create_cell(world_object)
-        cell.add(self, world_object)
+    def add_world_object_spawn(self, world_object_spawn):
+        cell = self.get_create_cell(world_object_spawn.location, world_object_spawn.map_)
+        cell.add_world_object_spawn(world_object_spawn)
+
+    def add_world_object(self, world_object, update_players=True):
+        cell: Cell = self.get_create_cell(world_object.location, world_object.map_)
+        cell.add_world_object(world_object)
 
         if world_object.get_type_id() == ObjectTypeIds.ID_PLAYER:
             affected_cells = list(self.get_surrounding_cells_by_object(world_object))
@@ -290,6 +296,15 @@ class GridManager:
                 return gameobject
         return None
 
+    def get_create_cell(self, vector, map_) -> Cell:
+        cell_key = GridManager.get_cell_key(vector.x, vector.y, map_)
+        cell = self.cells.get(cell_key)
+        if not cell:
+            min_x, min_y, max_x, max_y = GridManager.generate_coord_data(vector.x, vector.y)
+            cell = Cell(min_x, min_y, max_x, max_y, map_)
+            self.cells[cell.key] = cell
+        return cell
+
     @staticmethod
     def generate_coord_data(x, y):
         mod_x = x / CELL_SIZE
@@ -306,7 +321,6 @@ class GridManager:
     def get_cell_key(x, y, map_):
         min_x, min_y, max_x, max_y = GridManager.generate_coord_data(x, y)
         key = f'{round(min_x, 5)}:{round(min_y, 5)}:{round(max_x, 5)}:{round(max_y, 5)}:{map_}'
-
         return key
 
     def get_cells(self):
@@ -315,9 +329,7 @@ class GridManager:
     def update_creatures(self):
         now = time.time()
         for key in list(self.active_cell_keys):
-            cell = self.cells[key]
-            for guid, creature in list(cell.creatures.items()):
-                creature.update(now)
+            self.cells[key].update_creatures(now)
 
     def update_gameobjects(self):
         now = time.time()
@@ -332,95 +344,3 @@ class GridManager:
             cell = self.cells[key]
             for guid, corpse in list(cell.corpses.items()):
                 corpse.update(now)
-
-
-class Cell:
-    def __init__(self, min_x=0.0, min_y=0.0, max_x=0.0, max_y=0.0, map_=0.0, key=''):
-        self.min_x = min_x
-        self.min_y = min_y
-        self.max_x = max_x
-        self.max_y = max_y
-        self.map_ = map_
-        self.key = key
-        self.gameobjects = dict()
-        self.creatures = dict()
-        self.players = dict()
-        self.dynamic_objects = dict()
-        self.corpses = dict()
-
-        if not key:
-            self.key = f'{round(self.min_x, 5)}:{round(self.min_y, 5)}:{round(self.max_x, 5)}:{round(self.max_y, 5)}:{self.map_}'
-
-    def has_players(self):
-        return len(self.players) > 0
-
-    def contains(self, world_object=None, vector=None, map_=None):
-        if world_object:
-            vector = world_object.location
-            map_ = world_object.map_
-
-        if vector and map_:
-            return self.min_x <= round(vector.x, 5) <= self.max_x and \
-                   self.min_y <= round(vector.y, 5) <= self.max_y and \
-                   map_ == self.map_
-        return False
-
-    def add(self, grid_manager, world_object):
-        # Update world_object cell so the below messages affect the new cell surroundings.
-        world_object.current_cell = self.key
-
-        if world_object.get_type_id() == ObjectTypeIds.ID_PLAYER:
-            self.players[world_object.guid] = world_object
-        elif world_object.get_type_id() == ObjectTypeIds.ID_UNIT:
-            self.creatures[world_object.guid] = world_object
-        elif world_object.get_type_id() == ObjectTypeIds.ID_GAMEOBJECT:
-            self.gameobjects[world_object.guid] = world_object
-        elif world_object.get_type_id() == ObjectTypeIds.ID_DYNAMICOBJECT:
-            self.dynamic_objects[world_object.guid] = world_object
-        elif world_object.get_type_id() == ObjectTypeIds.ID_CORPSE:
-            self.corpses[world_object.guid] = world_object
-
-    # Make each player update its surroundings, adding, removing or updating world objects as needed.
-    def update_players(self, world_object=None, has_changes=False, has_inventory_changes=False):
-        for player in list(self.players.values()):
-            if world_object:
-                player.update_world_object_on_me(world_object, has_changes, has_inventory_changes)
-            else:
-                player.update_known_world_objects()
-
-    def remove(self, world_object):
-        if world_object.get_type_id() == ObjectTypeIds.ID_PLAYER:
-            self.players.pop(world_object.guid, None)
-        elif world_object.get_type_id() == ObjectTypeIds.ID_UNIT:
-            self.creatures.pop(world_object.guid, None)
-        elif world_object.get_type_id() == ObjectTypeIds.ID_GAMEOBJECT:
-            self.gameobjects.pop(world_object.guid, None)
-        elif world_object.get_type_id() == ObjectTypeIds.ID_DYNAMICOBJECT:
-            self.dynamic_objects.pop(world_object.guid, None)
-        elif world_object.get_type_id() == ObjectTypeIds.ID_CORPSE:
-            self.corpses.pop(world_object.guid, None)
-
-    def send_all(self, packet, source=None, exclude=None, use_ignore=False):
-        for guid, player_mgr in list(self.players.items()):
-            if player_mgr.online:
-                if source and player_mgr.guid == source.guid:
-                    continue
-                if exclude and player_mgr.guid in exclude:
-                    continue
-                if use_ignore and source and player_mgr.friends_manager.has_ignore(source.guid):
-                    continue
-
-                player_mgr.enqueue_packet(packet)
-
-    def send_all_in_range(self, packet, range_, source, include_self=True, exclude=None, use_ignore=False):
-        if range_ <= 0:
-            self.send_all(packet, source, exclude)
-        else:
-            for guid, player_mgr in list(self.players.items()):
-                if player_mgr.online and player_mgr.location.distance(source.location) <= range_:
-                    if not include_self and player_mgr.guid == source.guid:
-                        continue
-                    if use_ignore and player_mgr.friends_manager.has_ignore(source.guid):
-                        continue
-
-                    player_mgr.enqueue_packet(packet)
